Fixes issue #922 - Rename jndi directory to naming (introduces Piranha Naming)
