+++ extension/micro-core/pom.xml
@@ -74,7 +74,7 @@
                 
         <dependency>
             <groupId>cloud.piranha.naming</groupId>
-            <artifactId>piranha-naming-memory</artifactId>
+            <artifactId>piranha-naming-impl</artifactId>
             <version>${project.version}</version>
             <scope>provided</scope>
         </dependency>
