+++ extension/micro-servlet/pom.xml
@@ -113,10 +113,10 @@
             <scope>runtime</scope>
         </dependency>
         
-        <!-- Piranha Webapplication - web.xml support -->
+        <!-- Piranha Naming implementation -->
         <dependency>
             <groupId>cloud.piranha.naming</groupId>
-            <artifactId>piranha-naming-memory</artifactId>
+            <artifactId>piranha-naming-impl</artifactId>
             <version>${project.version}</version>
             <scope>runtime</scope>
         </dependency>
