+++ micro/src/test/java/cloud/piranha/micro/MicroPiranhaTest.java
@@ -51,7 +51,7 @@ public class MicroPiranhaTest {
     @Test
     @Disabled
     public void testStart() throws Exception {
-        System.setProperty("java.naming.factory.initial", "cloud.piranha.naming.memory.DefaultInitialContextFactory");
+        System.setProperty("java.naming.factory.initial", "cloud.piranha.naming.impl.DefaultInitialContextFactory");
         final MicroPiranha piranha = new MicroPiranha();
         piranha.configure(new String[]{});
         Thread thread = new Thread(piranha);
@@ -77,7 +77,7 @@ public class MicroPiranhaTest {
     @Test
     @Disabled
     public void testChangingPort() throws Exception {
-        System.setProperty("java.naming.factory.initial", "cloud.piranha.naming.memory.DefaultInitialContextFactory");
+        System.setProperty("java.naming.factory.initial", "cloud.piranha.naming.impl.DefaultInitialContextFactory");
         final MicroPiranha piranha = new MicroPiranha();
         piranha.configure(new String[]{"--port", "8088"});
         Thread thread = new Thread(piranha);
