+++ naming/impl/src/main/java/cloud/piranha/naming/impl/DefaultNamingEnumeration.java
@@ -25,7 +25,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package cloud.piranha.naming.memory;
+package cloud.piranha.naming.impl;
 
 import java.util.Collection;
 import java.util.Collections;
