+++ naming/impl/src/test/java/cloud/piranha/naming/impl/DefaultNameParserTest.java
@@ -25,7 +25,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package cloud.piranha.naming.memory;
+package cloud.piranha.naming.impl;
 
 import javax.naming.Name;
 import org.junit.jupiter.api.Test;
