+++ security/soteria/pom.xml
@@ -17,7 +17,7 @@
     <dependencies>
         <dependency>
             <groupId>cloud.piranha.naming</groupId>
-            <artifactId>piranha-naming-memory</artifactId>
+            <artifactId>piranha-naming-impl</artifactId>
             <version>${project.version}</version>
             <scope>provided</scope>
         </dependency>
