+++ security/soteria/src/main/java/cloud/piranha/security/soteria/DynamicInitialContextFactory.java
@@ -34,7 +34,7 @@ import javax.naming.Context;
 import javax.naming.NamingException;
 import javax.naming.spi.InitialContextFactory;
 
-import cloud.piranha.naming.memory.DefaultInitialContext;
+import cloud.piranha.naming.impl.DefaultInitialContext;
 
 /**
  * The default InitialContextFactory.
