+++ server/pom.xml
@@ -49,10 +49,10 @@
             <version>${project.version}</version>
             <scope>compile</scope>
         </dependency>
-        <!-- Piranha JNDI memory -->
+        <!-- Piranha Naming implementation -->
         <dependency>
             <groupId>cloud.piranha.naming</groupId>
-            <artifactId>piranha-naming-memory</artifactId>
+            <artifactId>piranha-naming-impl</artifactId>
             <version>${project.version}</version>
             <scope>compile</scope>
         </dependency>
