+++ test/jpa/src/test/java/cloud/piranha/test/jpa/JPATest.java
@@ -56,7 +56,7 @@ public class JPATest {
     @Test
     public void testIndexHtml() throws Exception {
         System.getProperties().put("java.naming.factory.initial", 
-                "cloud.piranha.naming.memory.DefaultInitialContextFactory");
+                "cloud.piranha.naming.impl.DefaultInitialContextFactory");
         InitialContext initialContext = new InitialContext();
         JDBCDataSource ds = new JDBCDataSource();
         ds.setUrl("jdbc:hsqldb:mem:demo");
