+++ test/openwebbeans/src/test/java/cloud/piranha/test/openwebbeans/HelloOpenWebBeansTest.java
@@ -53,7 +53,7 @@ public class HelloOpenWebBeansTest {
     @Test
     public void testIndexHtml() throws Exception {
         System.getProperties().put("java.naming.factory.initial",
-                "cloud.piranha.naming.memory.DefaultInitialContextFactory");
+                "cloud.piranha.naming.impl.DefaultInitialContextFactory");
         EmbeddedPiranha piranha = new EmbeddedPiranhaBuilder()
                 .directoryResource("src/main/webapp")
                 .aliasedDirectoryResource("target/classes", "/WEB-INF/classes")
