+++ test/smallrye-health/src/test/java/piranha/cloud/test/smallrye/health/SmallRyeHealthTest.java
@@ -53,7 +53,7 @@ public class SmallRyeHealthTest {
     @Test
     public void testHealth() throws Exception {
         System.getProperties().put("java.naming.factory.initial",
-                "cloud.piranha.naming.memory.DefaultInitialContextFactory");
+                "cloud.piranha.naming.impl.DefaultInitialContextFactory");
         EmbeddedPiranha piranha = new EmbeddedPiranhaBuilder()
                 .directoryResource("src/main/webapp")
                 .initializer(WeldInitializer.class.getName())
