+++ test/smallrye-health/src/test/java/piranha/cloud/test/smallrye/health/SmallRyeLivenessTest.java
@@ -53,7 +53,7 @@ public class SmallRyeLivenessTest {
     @Test
     public void testHealth() throws Exception {
         System.getProperties().put("java.naming.factory.initial", 
-                "cloud.piranha.naming.memory.DefaultInitialContextFactory");
+                "cloud.piranha.naming.impl.DefaultInitialContextFactory");
         EmbeddedPiranha piranha = new EmbeddedPiranhaBuilder()
                 .directoryResource("src/main/webapp")
                 .initializer(WeldInitializer.class.getName())
