+++ test/soteria/basic/src/test/java/cloud/piranha/test/soteria/basic/DynamicInitialContextFactory.java
@@ -32,7 +32,7 @@ import javax.naming.Context;
 import javax.naming.NamingException;
 import javax.naming.spi.InitialContextFactory;
 
-import cloud.piranha.naming.memory.DefaultInitialContext;
+import cloud.piranha.naming.impl.DefaultInitialContext;
 
 /**
  * The default InitialContextFactory.
