+++ test/vaadin/src/test/java/cloud/piranha/test/vaadin/HelloVaadinTest.java
@@ -51,7 +51,7 @@ public class HelloVaadinTest {
     @Test
     public void testIndexHtmlPage() throws Exception {
         System.getProperties().put("java.naming.factory.initial", 
-                "cloud.piranha.naming.memory.DefaultInitialContextFactory");
+                "cloud.piranha.naming.impl.DefaultInitialContextFactory");
         EmbeddedPiranha piranha = new EmbeddedPiranhaBuilder()
                 .directoryResource("src/main/webapp")
                 .servletMapping("Vaadin", "/*")
