+++ test/weld/src/test/java/cloud/piranha/test/weld/HelloWeldTest.java
@@ -54,7 +54,7 @@ public class HelloWeldTest {
     @Test
     public void testIndexHtml() throws Exception {
         System.getProperties().put("java.naming.factory.initial", 
-                "cloud.piranha.naming.memory.DefaultInitialContextFactory");
+                "cloud.piranha.naming.impl.DefaultInitialContextFactory");
         EmbeddedPiranha piranha = new EmbeddedPiranhaBuilder()
                 .directoryResource("src/main/webapp")
                 .aliasedDirectoryResource("target/classes", "/WEB-INF/classes")
